rename core script references
